PRSD-956: Moves certificate file name helper into PropertyComplianceJourneyHelper
