+++ src/test/kotlin/uk/gov/communities/prsdb/webapp/helpers/extensions/journeyExtensions/PropertyComplianceJourneyDataExtensionsTests.kt
@@ -4,7 +4,6 @@ import kotlinx.datetime.toKotlinLocalDate
 import org.junit.jupiter.api.BeforeEach
 import org.junit.jupiter.api.Named
 import org.junit.jupiter.api.Test
-import org.junit.jupiter.api.assertThrows
 import org.junit.jupiter.params.ParameterizedTest
 import org.junit.jupiter.params.provider.Arguments
 import org.junit.jupiter.params.provider.MethodSource
@@ -13,28 +12,24 @@ import org.mockito.Mockito.mockConstruction
 import org.mockito.kotlin.whenever
 import uk.gov.communities.prsdb.webapp.constants.enums.EicrExemptionReason
 import uk.gov.communities.prsdb.webapp.constants.enums.GasSafetyExemptionReason
-import uk.gov.communities.prsdb.webapp.forms.steps.PropertyComplianceStepId
 import uk.gov.communities.prsdb.webapp.helpers.DateTimeHelper
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyExtensions.Companion.getHasEICR
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyExtensions.Companion.getHasEicrExemption
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyExtensions.Companion.getHasGasSafetyCert
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyExtensions.Companion.getHasGasSafetyCertExemption
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyExtensions.Companion.getIsEicrExemptionReasonOther
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyExtensions.Companion.getIsEicrOutdated
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyExtensions.Companion.getIsGasSafetyCertOutdated
-import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyExtensions.Companion.getIsGasSafetyExemptionReasonOther
+import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getHasEICR
+import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getHasEicrExemption
+import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getHasGasSafetyCert
+import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getHasGasSafetyCertExemption
+import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getIsEicrExemptionReasonOther
+import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getIsEicrOutdated
+import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getIsGasSafetyCertOutdated
+import uk.gov.communities.prsdb.webapp.helpers.extensions.journeyExtensions.PropertyComplianceJourneyDataExtensions.Companion.getIsGasSafetyExemptionReasonOther
 import uk.gov.communities.prsdb.webapp.testHelpers.builders.JourneyDataBuilder
 import java.time.LocalDate
 import kotlin.test.assertEquals
 import kotlin.test.assertNull
 
-class PropertyComplianceJourneyExtensionsTests {
+class PropertyComplianceJourneyDataExtensionsTests {
     companion object {
         // currentDate is an arbitrary date
         private val currentDate = LocalDate.of(2020, 1, 5).toKotlinLocalDate()
-
-        private const val PROPERTY_OWNERSHIP_ID = 1L
-        private const val ORIGINAL_FILE_EXT = "png"
 
         @JvmStatic
         private fun provideGasSafetyCertIssueDates() =
@@ -64,19 +59,6 @@ class PropertyComplianceJourneyExtensionsTests {
             arrayOf(
                 Arguments.of(Named.of("other", EicrExemptionReason.OTHER), true),
                 Arguments.of(Named.of("not other", EicrExemptionReason.LIVE_IN_LANDLORD), false),
-            )
-
-        @JvmStatic
-        private fun provideFileUploadStepNames() =
-            arrayOf(
-                Arguments.of(
-                    Named.of(PropertyComplianceStepId.GasSafetyUpload.name, PropertyComplianceStepId.GasSafetyUpload.urlPathSegment),
-                    "property_${PROPERTY_OWNERSHIP_ID}_gas_safety_certificate.$ORIGINAL_FILE_EXT",
-                ),
-                Arguments.of(
-                    Named.of(PropertyComplianceStepId.EicrUpload.name, PropertyComplianceStepId.EicrUpload.urlPathSegment),
-                    "property_${PROPERTY_OWNERSHIP_ID}_eicr.$ORIGINAL_FILE_EXT",
-                ),
             )
     }
 
@@ -255,29 +237,5 @@ class PropertyComplianceJourneyExtensionsTests {
         val retrievedIsEicrExemptionReasonOther = testJourneyData.getIsEicrExemptionReasonOther()
 
         assertNull(retrievedIsEicrExemptionReasonOther)
-    }
-
-    @ParameterizedTest(name = "for the {0} step")
-    @MethodSource("provideFileUploadStepNames")
-    fun `getCertFilename returns the corresponding file name`(
-        stepName: String,
-        expectedFileName: String,
-    ) {
-        val originalFileName = "any-name.$ORIGINAL_FILE_EXT"
-
-        assertEquals(
-            expectedFileName,
-            PropertyComplianceJourneyExtensions.getCertFilename(PROPERTY_OWNERSHIP_ID, stepName, originalFileName),
-        )
-    }
-
-    @Test
-    fun `getCertFilename throws an IllegalStateException for invalid file upload step names`() {
-        val invalidStepName = "invalid-step"
-        val originalFileName = "any-name.$ORIGINAL_FILE_EXT"
-
-        assertThrows<IllegalStateException> {
-            PropertyComplianceJourneyExtensions.getCertFilename(PROPERTY_OWNERSHIP_ID, invalidStepName, originalFileName)
-        }
     }
 }
